+++ src/app/schools/byu/page.mdx
@@ -5,14 +5,16 @@ import { SplitPane } from '@/components/atoms/SplitPane';
 import { PhotoViewer } from '@/components/molecules/PhotoViewer';
 import { BYU_PAGE_META } from '@/data/links';
 import { SCHOOL_INFO } from '@/data/school';
-import archesNationalParkImage from '@/features/byu/assets/arches-national-park.jpg';
-import byuFootballImage from '@/features/byu/assets/byu-football.jpg';
-import byuLargeClass from '@/features/byu/assets/byu-large-class.jpg';
-import byuLogoImage from '@/features/byu/assets/byu-logo.png';
-import canyonLandsNationalParkImage from '@/features/byu/assets/canyonlands-national-park.jpg';
-import saltFlatsImage from '@/features/byu/assets/salt-flats.jpg';
-import skiResortImage from '@/features/byu/assets/ski-resort.jpg';
-import utahStateCapitalBuilding from '@/features/byu/assets/utah-state-capital.jpg';
+import {
+  archesNationalParkImage,
+  byuFootballImage,
+  byuLargeClass,
+  byuLogoImage,
+  canyonLandsNationalParkImage,
+  saltFlatsImage,
+  skiResortImage,
+  utahStateCapitalBuilding,
+} from '@/features/byu';
 
 export const metadata = {
   title: BYU_PAGE_META.title,
