+++ src/app/schools/byuh/page.mdx
@@ -5,15 +5,17 @@ import { SplitPane } from '@/components/atoms/SplitPane';
 import { PhotoViewer } from '@/components/molecules/PhotoViewer';
 import { BYUH_PAGE_META } from '@/data/links';
 import { SCHOOL_INFO } from '@/data/school';
-import byuhClassImage from '@/features/byuh/assets/byuh-class.jpg';
-import byuhCultureNightImage from '@/features/byuh/assets/byuh-culturenight.jpg';
-import byuhDressAndGroomingImage from '@/features/byuh/assets/byuh-dress-and-grooming.jpg';
-import byuhFoodFestaImage from '@/features/byuh/assets/byuh-food-festa.jpg';
-import byuhHousing from '@/features/byuh/assets/byuh-housing.jpg';
-import byuhBbqImage from '@/features/byuh/assets/byuh-japan-student-chapter-bbq.jpg';
-import byuhResumeWorkshopImage from '@/features/byuh/assets/byuh-japan-student-chapter-resume-workshop.jpg';
-import byuhLogoImage from '@/features/byuh/assets/byuh-logo.png';
-import byuhLuauImage from '@/features/byuh/assets/pcc-luau.jpg';
+import {
+  byuhClassImage,
+  byuhCultureNightImage,
+  byuhDressAndGroomingImage,
+  byuhFoodFestaImage,
+  byuhHousing,
+  byuhBbqImage,
+  byuhResumeWorkshopImage,
+  byuhLogoImage,
+  byuhLuauImage,
+} from '@/features/byuh';
 
 export const metadata = {
   title: BYUH_PAGE_META.title,
