+++ src/app/schools/byui/page.mdx
@@ -5,10 +5,12 @@ import { SplitPane } from '@/components/atoms/SplitPane';
 import { PhotoViewer } from '@/components/molecules/PhotoViewer';
 import { BYUI_PAGE_META } from '@/data/links';
 import { SCHOOL_INFO } from '@/data/school';
-import byuiClassImage from '@/features/byui/assets/byui-class.png';
-import byuiFitnessImage from '@/features/byui/assets/byui-fitness.png';
-import byuiHousingImage from '@/features/byui/assets/byui-housing.png';
-import byuiLogoImage from '@/features/byui/assets/byui-logo.png';
+import {
+  byuiClassImage,
+  byuiFitnessImage,
+  byuiHousingImage,
+  byuiLogoImage,
+} from '@/features/byui';
 
 export const metadata = {
   title: BYUI_PAGE_META.title,
