+++ src/app/tips/apply/page.mdx
@@ -7,7 +7,7 @@ import {
 } from '@/data/application-deadline';
 import { ENGLISH_REQUIREMENT_BY_SCHOOL } from '@/data/english-requirement';
 import { APPLY_PAGE_META } from '@/data/links';
-import { EnglishRequirementBySchoolTable } from '@/features/apply/components/EnglishRequirementBySchoolTable';
+import { EnglishRequirementBySchoolTable } from '@/features/apply';
 
 export const metadata = {
   title: APPLY_PAGE_META.title,
