+++ src/app/tips/career/page.mdx
@@ -2,7 +2,7 @@ import { Footnote } from '@/components/atoms/Footnote';
 import { UlCol } from '@/components/atoms/UlCol';
 import { CAREER_PAGE_META } from '@/data/links';
 import { MDX } from '@/mdx-components';
-import { IncomeBarChart } from '@/features/career/components/IncomeBarChart';
+import { IncomeBarChart } from '@/features/career';
 
 export const metadata = {
   title: CAREER_PAGE_META.title,
