+++ src/app/tips/compare/page.mdx
@@ -4,10 +4,12 @@ import { ENGLISH_REQUIREMENT_BY_EXAM } from '@/data/english-requirement';
 import { ExrateAnnotation } from '@/features/forex';
 import { MDX } from '@/mdx-components';
 import { CurrencySwitchControl, LDSSwitchControl } from '@/features/stats';
-import { CostTable } from '@/features/compare/components/CostTable';
-import { EnglishRequirementByExamTable } from '@/features/compare/components/EnglishRequirementByExamTable';
-import { SchoolDataTable } from '@/features/compare/components/SchoolDataTable';
-import { TuitionBarChart } from '@/features/compare/components/TuitionBarChart';
+import {
+  CostTable,
+  EnglishRequirementByExamTable,
+  SchoolDataTable,
+  TuitionBarChart,
+} from '@/features/compare';
 
 
 # BYU、BYUH、BYUIの比較
