+++ src/app/tips/internship/page.mdx
@@ -2,7 +2,7 @@ import { Footnote } from '@/components/atoms/Footnote';
 import { UlCol } from '@/components/atoms/UlCol';
 import { INTERNSHIP_PAGE_META } from '@/data/links';
 import { MDX } from '@/mdx-components';
-import { InternSalaryPieChart } from '@/features/internship/components/InternSalaryPieChart';
+import { InternSalaryPieChart } from '@/features/internship';
 
 export const metadata = {
   title: INTERNSHIP_PAGE_META.title,
