+++ src/components/atoms/NextImage/NextImage.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj, StoryFn } from '@storybook/react';
-import skiResortImage from '@/features/byu/assets/ski-resort.jpg';
+import { skiResortImage } from '@/features/byu';
 import { NextImage } from './NextImage';
 
 const meta = {
