+++ src/components/molecules/SchoolTh/SchoolTh.tsx
@@ -1,7 +1,7 @@
 import { NextImage } from '@/components/atoms/NextImage';
-import byuLogoImage from '@/features/byu/assets/byu-logo.png';
-import byuhLogoImage from '@/features/byuh/assets/byuh-logo.png';
-import byuiLogoImage from '@/features/byui/assets/byui-logo.png';
+import { byuLogoImage } from '@/features/byu';
+import { byuhLogoImage } from '@/features/byuh';
+import { byuiLogoImage } from '@/features/byui';
 
 type ISchoolTh = () => JSX.Element;
