+++ src/features/byui/index.ts
@@ -1,0 +1,4 @@
+export { default as byuiClassImage } from './assets/byui-class.png';
+export { default as byuiFitnessImage } from './assets/byui-fitness.png';
+export { default as byuiHousingImage } from './assets/byui-housing.png';
+export { default as byuiLogoImage } from './assets/byui-logo.png';
