+++ src/features/compare/index.ts
@@ -1,0 +1,4 @@
+export * from './components/CostTable';
+export * from './components/EnglishRequirementByExamTable';
+export * from './components/SchoolDataTable';
+export * from './components/TuitionBarChart';
